make IdVal more generic
